+++ app/src/Diskora.App/VirtualDiskWindow.xaml.cs
@@ -1,6 +1,7 @@
 using System.Windows;
 using Diskora.App.ViewModels;
 using Diskora.Core.Services;
+using Diskora.Data;
 
 namespace Diskora.App;
 
@@ -9,6 +10,6 @@ public partial class VirtualDiskWindow : Window
     public VirtualDiskWindow(string path)
     {
         InitializeComponent();
-        DataContext = new VirtualDiskViewModel(new VirtualDiskService(), path);
+        DataContext = new VirtualDiskViewModel(new VirtualDiskService(new SqliteVirtualDiskAttachmentRegistry()), path);
     }
 }